requirements.txt
python_telegram_bot
requests
python_upwork
freelancersdk
pymongo
celery
redis
